Implement serial_initialize() and serial_write().
These work with QEMU... but they also work under the following strange
conditions:
  * Neglect to call serial_initialize().
  * Make serial_can_transmit() return true unconditionally.

QEMU is insensitive to these issues, but a physical UART would might
have trouble.  Since we have already demonstrated that QEMU is
insensitive to some potentially devastating bugs, I am not confident
that this implementation is correct.  It should be tested on real
hardware at some point!