Implement the a boot allocator.
And, use the boot allocator to allocate space for the memory catalog.
Now, the linker doesn't need to know anything about what we use the
pre-allocated space for.  It just knows that it has a boot heap, and
it's up to us to specify a big enough size for it!