Begin implementing single-type table generation.
This will allow us to lookup the file/line/function of each instruction.

This commit is pretty half-baked.  It turns out that this branch
should use the non-boot allocator.  That hasn't been implemented, yet.
So, I'm committing this half-baked work so I can more easily work on
the non-boot allocator.