Add a dummy COM1 interrupt handler.
For now, the handler just prints the received character.  In an
upcoming pull request, we will make it feed some a system input buffer.

As part of this change, I implemented serial-read functionality.  (It
was very simple-- I had already defined the relevant ports and
bitfields in a previous pull request.)